Reviewed and will Synchronize this Time o:)
 handed in the assignment a week  ago and now I reviewed it and I will
be submitting again. There are many things I learned from the feedback
about the organization of HTML and CSS documents.
There is a Python notes document attached that needs quite a bit of
work and I will probably reset it to work with the new formatting and
to give each section appropriate class names to get easy and quick
styling. I was experimenting with the different things CSS enables you
to do and I added a few design touches to the document! :)
Side Note: I submitted it last time without synchronizing my GitHub so
I guess I will need to practice using GitHub a bit more and maybe take
that course on it.